+++ src/main/kotlin/com/rtomyj/skc/dao/BanListDao.kt
@@ -1,4 +1,4 @@
-package com.rtomyj.skc.find.banlist.dao
+package com.rtomyj.skc.dao
 
 import com.rtomyj.skc.exception.SKCException
 import com.rtomyj.skc.model.BanListDates
