+++ src/main/kotlin/com/rtomyj/skc/stats/StatsService.kt
@@ -1,10 +1,10 @@
 package com.rtomyj.skc.stats
 
+import com.rtomyj.skc.dao.StatsDao
 import com.rtomyj.skc.exception.ErrorType
 import com.rtomyj.skc.exception.SKCException
 import com.rtomyj.skc.model.DatabaseStats
 import com.rtomyj.skc.model.MonsterTypeStats
-import com.rtomyj.skc.stats.dao.StatsDao
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.beans.factory.annotation.Qualifier
