+++ src/test/kotlin/com/rtomyj/skc/dao/BanListJDBCDaoTest.kt
@@ -1,8 +1,9 @@
-package com.rtomyj.skc.find.banlist.dao
+package com.rtomyj.skc.dao
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.rtomyj.skc.config.DateConfig
 import com.rtomyj.skc.constant.TestObjects
+import com.rtomyj.skc.dao.BanListJDBCDao
 import com.rtomyj.skc.model.Card
 import com.rtomyj.skc.util.enumeration.BanListCardStatus
 import org.junit.jupiter.api.Assertions
