+++ src/test/kotlin/com/rtomyj/skc/find/banlist/service/BannedCardsServiceTest.kt
@@ -3,9 +3,9 @@ package com.rtomyj.skc.find.banlist.service
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.nhaarman.mockito_kotlin.eq
 import com.rtomyj.skc.constant.TestConstants
+import com.rtomyj.skc.dao.BanListDao
 import com.rtomyj.skc.exception.ErrorType
 import com.rtomyj.skc.exception.SKCException
-import com.rtomyj.skc.find.banlist.dao.BanListDao
 import com.rtomyj.skc.model.BanListInstance
 import com.rtomyj.skc.model.BanListNewContent
 import com.rtomyj.skc.model.BanListRemovedContent
