+++ src/test/kotlin/com/rtomyj/skc/find/banlist/service/DiffServiceTest.kt
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.nhaarman.mockito_kotlin.any
 import com.nhaarman.mockito_kotlin.eq
 import com.rtomyj.skc.constant.TestConstants
+import com.rtomyj.skc.dao.BanListDao
 import com.rtomyj.skc.exception.SKCException
-import com.rtomyj.skc.find.banlist.dao.BanListDao
 import com.rtomyj.skc.model.BanListNewContent
 import com.rtomyj.skc.model.BanListRemovedContent
 import com.rtomyj.skc.util.enumeration.BanListCardStatus
