+++ src/test/kotlin/com/rtomyj/skc/find/card/CardServiceTest.kt
@@ -1,13 +1,13 @@
 package com.rtomyj.skc.find.card
 
 import com.nhaarman.mockito_kotlin.eq
-import com.rtomyj.skc.browse.product.dao.ProductDao
 import com.rtomyj.skc.constant.TestConstants
 import com.rtomyj.skc.constant.TestObjects
+import com.rtomyj.skc.dao.BanListDao
+import com.rtomyj.skc.dao.Dao
+import com.rtomyj.skc.dao.ProductDao
 import com.rtomyj.skc.exception.ErrorType
 import com.rtomyj.skc.exception.SKCException
-import com.rtomyj.skc.find.banlist.dao.BanListDao
-import com.rtomyj.skc.find.card.dao.Dao
 import com.rtomyj.skc.model.Card
 import com.rtomyj.skc.skcsuggestionengine.traffic.TrafficService
 import com.rtomyj.skc.util.enumeration.TrafficResourceType
