+++ src/test/kotlin/com/rtomyj/skc/status/StatusControllerTest.kt
@@ -1,12 +1,12 @@
 package com.rtomyj.skc.status
 
+import com.rtomyj.skc.dao.StatusDao
 import com.rtomyj.skc.exception.ErrorType
 import com.rtomyj.skc.exception.SKCException
 import com.rtomyj.skc.model.DownstreamStatus
 import com.rtomyj.skc.model.SuggestionEngineDownstreamStatus
 import com.rtomyj.skc.model.SuggestionEngineStatus
 import com.rtomyj.skc.skcsuggestionengine.status.SuggestionEngineStatusService
-import com.rtomyj.skc.status.dao.StatusDao
 import org.hamcrest.Matchers.`is`
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Tag
